build: Add support for faking platform security patch level
SafetyNet will fail if the fake build fingerprints don't match the
platform security patch level, which is a problem for devices that are
no longer supported by the manufacturer.

To fix this issue, allow legacy products to define a fake security patch
level from their last official update and set the standard patch level
property to that. The real patch level is moved to another custom
property that is shown to the user, e.g. in Settings -> About.

Note that this requires the "core: Use real security patch level
property" commit in frameworks/base for users to see the correct patch
level in Settings.

Also includes:

Author: Danny Lin <danny@kdrag0n.dev>
Date:   Mon Nov 9 08:49:20 2020 -0800

    build: Swap fake and real platform security patch levels

    Instead of defining a fake level to use for the property, use the
    fake level for everything and define a new level to use as the real
    one. This is necessary because Qualcomm's keymaster checks whether
    the platform and boot image security patch levels match.

    Change-Id: I173cf5c8d7cdde6e70967351a58da2222642d358

Author: Chirayu Desai <chirayudesai1@gmail.com>
Date:   Tue Sep 20 04:16:28 2022 +0530

    fixup! build: Add support for faking platform security patch level

    * The timestamp is used as AVB rollback index, so to avoid a
      mismatch between the security patch level and the timestamp,
      allow overriding it.

    Change-Id: I600ce5bf7fc5fdf0a9ab939628e60aae55df20c3

Change-Id: Id09e1b52697e6791987e388551e6f7a6558775ba