build: Add support for faking platform security patch level
SafetyNet will fail if the fake build fingerprints don't match the
platform security patch level, which is a problem for devices that are
no longer supported by the manufacturer.

To fix this issue, allow legacy products to override the platform
security patch level to the one from their last official update.
The real patch level is moved to another custom property that is shown
to the user, e.g. in Settings -> About.

Instead of defining a fake level to use for the property, use the
fake level for everything and define a new level to use as the real
one. This is necessary because Qualcomm's keymaster checks whether
the platform and boot image security patch levels match.

Note that this requires the "core: Use real security patch level
property" commit in frameworks/base for users to see the correct patch
level in Settings.

Also allow overriding the the timestamp, which is used as AVB rollback
index on some devices, to avoid a mismatch between the security patch
level and the timestamp.

Co-authored-by: Chirayu Desai <chirayudesai1@gmail.com>
Co-authored-by: Michael Bestas <mkbestas@gmail.com>
Change-Id: Id09e1b52697e6791987e388551e6f7a6558775ba