releasetools: Use apksigner for most APK and APEX signing
Instead of using the signapk tool, use the apksigner tool, except for
whole-file signing which will continue to be performed by signapk.
This change expects support for the -a (alignment) argument to have
been added to apksigner, as it is required for APEX container signing.

This behavior can be overridden with USE_APKSIGNER=n.

Issue: calyxos#3430
Change-Id: Icf645f9a0a9b928a424776e7069ab2fa0f3c6ea3