Add option to use a sign command intermediary
Add the --sign_command_intermediary option. Any commands involving
signapk, apksigner, avbtool, or custom payload signing will be passed
to the intermediary instead as arguments. This allows the intermediary
to perform any necessary setup, teardown, auditing, error handling, etc.
The intermediary is responsible for actually executing the command.

Note that if --payload_signer is specified, it will also be affected
and passed to the intermediary.

Issue: calyxos#3431
Change-Id: I862d5a15beade93f2cfad55a5fce49533f49d5a1